vtworker: Move execution state into "executor" struct.
This way, we don't have to pass all required parameters between functions all the time.
